unification of saleStatus
